requirements.txt
Package             Version
------------------- ---------
asgiref             3.5.1
certifi             2021.10.8
cffi                1.15.0
charset-normalizer  2.0.12
coreapi             2.3.3
coreschema          0.0.4
cryptography        37.0.2
defusedxml          0.7.1
dj-rest-auth        2.2.4
Django              4.0.4
django-allauth      0.50.0
django-rest-auth    0.9.5
django-rest-swagger 2.2.0
djangorestframework 3.13.1
idna                3.3
itypes              1.2.0
Jinja2              3.1.2
MarkupSafe          2.1.1
oauthlib            3.2.0
openapi-codec       1.3.2
pip                 21.3.1
pycparser           2.21
PyJWT               2.3.0
python3-openid      3.2.0
pytz                2022.1
PyYAML              6.0
requests            2.27.1
requests-oauthlib   1.3.1
setuptools          60.5.0
simplejson          3.17.6
six                 1.16.0
sqlparse            0.4.2
uritemplate         4.1.1
urllib3             1.26.9
wheel               0.37.1
